# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  firs_models.FirsTransactionVatPayload:
    properties:
      agentTin:
        type: string
      baseAmount:
        type: string
      beneficiaryTin:
        type: string
      currency:
        type: integer
      itemDescription:
        type: string
      otherTaxes:
        type: string
      totalAmount:
        type: string
      transDate:
        type: string
      vatCalculated:
        type: string
      vatRate:
        type: string
      vatStatus:
        type: integer
      vendorTransactionId:
        type: string
    type: object
...
  firs_models.VerifyTinData:
    properties:
      tin:
        type: string
    required:
    - tin
    type: object
  models.AccountingPlatformConfigAuth:
    properties:
      api_key:
        type: string
      api_secret:
        type: string
      auth_token:
        type: string
      hmac_secret:
        type: string
      org_id:
        type: string
    type: object
  models.CreateUserRequestModel:
    properties:
      email:
        type: string
      name:
        maxLength: 250
        minLength: 2
        type: string
      password:
        minLength: 6
        type: string
      platform_configs:
        $ref: '#/definitions/models.PlatformConfigsAuth'
    required:
    - email
    - name
    - password
    type: object
  models.LoginRequestModel:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  models.PlatformConfigsAuth:
    additionalProperties:
      $ref: '#/definitions/models.AccountingPlatformConfigAuth'
    type: object
  models.Response:
    properties:
      data: {}
      error: {}
      extra: {}
      message:
        type: string
      name:
        type: string
      pagination: {}
      status:
        type: string
      status_code:
        type: integer
    type: object
info:
  contact: {}
  description: This is the e-invoicing service API documentation.
  termsOfService: http://swagger.io/terms/
  title: Gention E-invoicing Service API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login to the system
      parameters:
      - description: Login request payload
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequestModel'
      produces:
      - application/json
      responses:
        "200":
          description: Login successfully
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad request, validation failed
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "422":
          description: Unprocessable entity
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Login
      tags:
      - Auth
  /auth/logout:
    get:
      consumes:
      - application/json
      description: Logout from the system
      produces:
      - application/json
      responses:
        "200":
          description: user logout successfully
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad request, validation failed
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "422":
          description: Unprocessable entity
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Logout
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Onboard to the system
      parameters:
      - description: Login request payload
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.CreateUserRequestModel'
      produces:
      - application/json
      responses:
        "200":
          description: Registered successfully
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad request, validation failed
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "422":
          description: Unprocessable entity
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Register
      tags:
      - Auth
...
  /entity:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of entities
      parameters:
      - in: query
        name: page
        type: integer
      - in: query
        name: reference
        type: string
      - in: query
        name: size
        type: integer
      - in: query
        name: sortBy
        type: string
      - in: query
        name: sortDirectionDesc
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Entities retrieved successfully
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Get Entities
      tags:
      - Entity
  /entity/{entity_id}:
    get:
      consumes:
      - application/json
      description: Retrieve details of a specific entity using its ID
      parameters:
      - description: Entity ID
        in: path
        name: entity_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Entity retrieved successfully
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Entity not found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Get Entity by ID
      tags:
      - Entity
  /entity/vat-payment:
    post:
      consumes:
      - application/json
      description: Submit VAT payment details to FIRS
      parameters:
      - description: VAT payment request payload
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/firs_models.FirsTransactionVatPayload'
      produces:
      - application/json
      responses:
        "200":
          description: VAT payment processed successfully
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "422":
          description: Validation failed
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Post VAT Payment
      tags:
      - Entity
  /entity/verify-tin:
    post:
      consumes:
      - application/json
      description: Verify a taxpayer identification number (TIN)
      parameters:
      - description: TIN verification request
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/firs_models.VerifyTinData'
      produces:
      - application/json
      responses:
        "200":
          description: TIN verified successfully
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "422":
          description: Validation failed
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Verify TIN
      tags:
      - Entity
...
  /zoho/auth/access-token:
    get:
      consumes:
      - application/json
      description: Exchange an authorization code for a Zoho access token and save
        it to the database.
      parameters:
      - description: Authorization Code returned from Zoho OAuth
        in: query
        name: code
        required: true
        type: string
      - description: Zoho Organization ID
        in: query
        name: organisation_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Token generated successfully
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad request (missing code or organisation_id)
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Get Zoho Access Token
      tags:
      - Zoho
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"